requirements.txt
Django==4.2
gunicorn==20.1.0
psycopg2-binary==2.9.10  # For PostgreSQL if you use it (or remove if using SQLite)
dj-database-url==0.5.0
python-dotenv==0.19.2
python-environ